requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
astunparse==1.6.3
audioread==3.0.1
blis==1.3.0
booknlp==1.0.8
catalogue==2.0.10
certifi==2025.7.9
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chatterbox-tts==0.1.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
conformer==0.3.2
cymem==2.0.11
decorator==5.2.1
diffusers==0.29.0
distlib==0.3.9
einops==0.8.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.33.4
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
murmurhash==1.0.13
namex==0.1.0
networkx==3.5
nodeenv==1.9.1
numba==0.61.2
numpy==2.1.3
onnx==1.18.0
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
pre_commit==4.2.0
preshed==3.0.10
protobuf==5.29.5
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
PyPDF2==3.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
resemble-perth==1.0.1
rich==14.0.0
s3tokenizer==0.2.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
shellingham==1.5.4
simpleaudio==1.0.4
six==1.17.0
smart_open==7.3.0.post1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.46.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
Unidecode==1.4.0
urllib3==2.5.0
virtualenv==20.31.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
